feat: int64 support + cfMesh-first Fine tier for large meshes
Python int64:
- PolyMeshWriter: int64 for tet connectivity
- file_reader.py: int64 for face indices
- repair.py: int64 for pymeshfix input
- NativeMeshChecker: already int64

Large mesh support:
- Fine tier: cfMesh first (no blockMesh → no int32 limit), snappy fallback
- Max background cells: fine=200M (was 50M)
- get_openfoam_label_size(): auto-detect label=32 or label=64
- Note: OpenFOAM label=32 limits blockMesh to ~2B cells
  cfMesh has no such limit (own background mesh)

To use OpenFOAM with >2B cells: recompile with WM_LABEL_SIZE=64

458 tests pass.

+++ core/analyzer/file_reader.py
@@ -130,7 +130,7 @@ def _load_via_cadquery(path: Path, fmt: str) -> trimesh.Trimesh:
         raise ValueError(f"cadquery 테셀레이션 결과가 비어 있습니다: {path}")
 
     vertices_np = np.array([(v.x, v.y, v.z) for v in vertices_cq], dtype=np.float64)
-    faces_np = np.array(faces_list, dtype=np.int32)
+    faces_np = np.array(faces_list, dtype=np.int64)
 
     mesh = trimesh.Trimesh(vertices=vertices_np, faces=faces_np, process=True)
